fix: - added sonar - fixed error messages (key) - added no auth
