+++ frontend/src/model/AppUser.tsx
@@ -2,7 +2,6 @@ type AppUser = {
     id: string;
     username: string;
     password: string;
-    role: string;
 }
 
 export default AppUser;
