+++ sonar-project.properties
@@ -1,0 +1,4 @@
+sonar.projectKey=adrianschulze_neuefische-capstone-project-adrian-schulze-frontend
+sonar.organization=adrianschulze
+sonar.projectName=adrianschulze_neuefische-capstone-project-adrian-schulze-frontend
+sonar.sources=./frontend/src
